chore(scanjob): move legacy serializers to a dedicated module
