+++ quipucords/api/serializers.py
@@ -19,9 +19,9 @@ from api.inspectresult.serializer import (
 )
 from api.report.serializer_v1 import ReportUploadSerializer
 from api.scan.serializer import ScanSerializer
-from api.scanjob.serializer import (
+from api.scanjob.serializer import ScanJobSerializer
+from api.scanjob.serializer_v1 import (
     ScanJobSerializerV1,
-    ScanJobSerializerV2,
     SimpleScanJobSerializer,
     SourceField,
 )
